libstagefright: Recover when we start to thrash the cache
When you have multiple streams and we must "seek", there is a
magic parameter hard-coded to 256KB.  If the streams are within
that many bytes of each other then everything is okay.  For
example,

Stream A: asks for X + 200KB
Stream B: asks for X

then we either seek to X + 200KB - 256KB or X - 256KB (depending
on which seek happens first) and in either case there will be
1 seek (aka audio/video glitch) and things are good.

However if we exceed the magic constant then things degenerate.

If stream A asks for X + 260KB then we will seek to X + 4KB but
stream B then reads @ X causing it to seek to X - 256KB.  If
stream A then reads @ X + 270KB (because it's moving forward) we
will then seek to that location and bounce back and forth with
audio and/or video going to hell until we happen to buffer
enough data to cover the window.

Work around this case by forcing buffering to happen if we
are trying to read data that is in the window that would have
been loaded if we had successfully prefetched at least the
lowwater threshold amount of data (but we just haven't had a
chance to actually prefetch it).

Change-Id: I356f28ea1cfdc9cc11f56e1bcd3002a20feffd1c

Patchset: 2
http://review.cyanogenmod.org/#/c/107669/